+++ edge/frontend/src/console.ts
@@ -1,9 +1,0 @@
-import { initializePreviews } from "./preview";
-import { initializeLiveDashboard } from "./live";
-import { initializeSemanticForms } from "./semantic";
-import { initializeShell } from "./shell";
-
-initializeShell();
-initializeLiveDashboard();
-initializeSemanticForms();
-initializePreviews();